Related work items: #15088

+++ dotnet/corecomponents/net.atos.daf.ct2.reports/entity/FleetOverviewResult.cs
@@ -13,7 +13,7 @@ namespace net.atos.daf.ct2.reports.entity
         public long Lcts_EndTimeStamp { get; set; }
         public string Lcts_Driver1Id { get; set; }
         public int Lcts_TripDistance { get; set; }
-        public int Lcts_DrivingTime { get; set; }
+        public long Lcts_DrivingTime { get; set; }
         public int Lcts_FuelConsumption { get; set; }
         public string Lcts_VehicleDrivingStatusType { get; set; }
         public long Lcts_OdometerVal { get; set; }
